+++ CHANGELOG.md
@@ -3,6 +3,7 @@ CHANGELOG
 
 ## HEAD (Unreleased)
 * Add constants for Aurora-based RDS Cluster `EngineType` and `EngineMode`
+* Add constants for EC2 Seurity Group Protocols
 
 ---