+++ CHANGELOG.md
@@ -4,6 +4,7 @@ CHANGELOG
 ## HEAD (Unreleased)
 * Add constants for Aurora-based RDS Cluster `EngineType` and `EngineMode`
 * Add constants for EC2 Seurity Group Protocols
+* Add constants for ALB IpAddressType and LoadBalancer type
 
 ---